res/requirements.txt
Flask
Redis
kafka-python
numpy
